Refactor duplicate code for the sequelize filters to utility method
Also changes some `var` to `const` at the top-level requires and remove some unnecessary & commented code.
