+++ src/models/Submission.js
@@ -1,5 +1,6 @@
 const config = require('config');
-var Sequelize = require('sequelize');
+const Sequelize = require('sequelize');
+const getSequelizeConditionsForFilters = require('./../util/getSequelizeConditionsForFilters');
 
 module.exports = function( db, sequelize, DataTypes ) {
 	var Submission = sequelize.define('submission', {
@@ -45,24 +46,6 @@ module.exports = function( db, sequelize, DataTypes ) {
 						value = JSON.parse(value);
 					}
 				} catch(err) {}
-				let newValue = {};
-
-				value = value;
-
-				/*
-				const configExtraData = [];
-				if (configExtraData) {
-					Object.keys(configExtraData).forEach((key) => {
-						if (configExtraData[key].allowNull === false && (typeof value[key] === 'undefined' || value[key] === '')) { // TODO: dit wordt niet gechecked als je het veld helemaal niet meestuurt
-							// zie validExtraData hieronder
-							// throw db.sequelize.ValidationError(`${key} is niet ingevuld`);
-						}
-						if (value[key] && configExtraData[key].values.indexOf(value[key]) != -1) { // TODO: alles is nu enum, maar dit is natuurlijk veel te simpel
-							newValue[key] = value[key];
-						}
-					});
-				}
-				*/
 
 				this.setDataValue('submittedData', JSON.stringify(value));
 			}
@@ -92,10 +75,6 @@ module.exports = function( db, sequelize, DataTypes ) {
                 };
             },
             filter:    function (filtersInclude, filtersExclude) {
-                let conditions = {
-                    [Sequelize.Op.and]: []
-                };
-                
                 const filterKeys = [
                     {
                         'key': 'id'
@@ -107,56 +86,8 @@ module.exports = function( db, sequelize, DataTypes ) {
                         'key': 'formId'
                     },
                 ];
-                
-                filterKeys.forEach((filter, i) => {
-                    //first add include filters
-                    if (filtersInclude) {
-                        let filterValue = filtersInclude[filter.key];
-                        
-                        if (filtersInclude[filter.key]) {
-                            if (filter.extraData) {
-                                filterValue = Array.isArray(filterValue) ? filterValue : [filterValue];
-                                
-                                const escapedKey = sequelize.escape(`$.${filter.key}`);
-                                filterValue.forEach((value, key) => {
-                                    const escapedValue = sequelize.escape(value);
-                                    conditions[Sequelize.Op.and].push({
-                                        [Sequelize.Op.and]: sequelize.literal(`extraData->${escapedKey}=${escapedValue}`)
-                                    });
-                                });
-                                
-                            } else {
-                                conditions[Sequelize.Op.and].push({
-                                    [filter.key]: filterValue
-                                });
-                            }
-                        }
-                    }
-                    
-                    //add exclude filters
-                    if (filtersExclude) {
-                        let excludeFilterValue = filtersExclude[filter.key];
-                        
-                        if (excludeFilterValue) {
-                            if (filter.extraData) {
-                                excludeFilterValue = Array.isArray(excludeFilterValue) ? excludeFilterValue : [excludeFilterValue];
-                                
-                                //filter out multiple conditions
-                                const escapedKey = sequelize.escape(`$.${filter.key}`);
-                                excludeFilterValue.forEach((value, key) => {
-                                    const escapedValue = sequelize.escape(value);
-                                    conditions[Sequelize.Op.and].push({
-                                        [Sequelize.Op.and]: sequelize.literal(`extraData->${escapedKey}!=${escapedValue}`)
-                                    });
-                                })
-                            }
-                        }
-                    }
-                });
-                
-                return {
-                    where: sequelize.and(conditions)
-                }
+    
+                return getSequelizeConditionsForFilters(filterKeys, filtersInclude, sequelize, filtersExclude);
             },
 		};
 	}
